Добавлено отображение списка отраслей в соответствующем фильтре
