--- app/src/main/java/ru/practicum/android/diploma/filter/industry/domain/models/Industry.kt
@@ -1,6 +1,0 @@
-package ru.practicum.android.diploma.filter.industry.domain.models
-
-data class Industry(
-    val id: Int,
-    val name: String
-)
