--- app/src/main/java/ru/practicum/android/diploma/filter/industry/ui/IndustryFilterScreen.kt
@@ -1,177 +1,0 @@
-@file:Suppress("MagicNumber")
-
-package ru.practicum.android.diploma.filter.industry.ui
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.selection.selectableGroup
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.KeyboardActions
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.OutlinedTextFieldDefaults
-import androidx.compose.material3.RadioButton
-import androidx.compose.material3.RadioButtonDefaults
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalFocusManager
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import ru.practicum.android.diploma.R
-import ru.practicum.android.diploma.filter.industry.domain.models.Industry
-import ru.practicum.android.diploma.ui.navigation.ActionBack
-import ru.practicum.android.diploma.ui.navigation.AppBarTop
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun IndustryFilterScreen(
-    industries: List<Industry> = getListIndustry(),
-    onBack: () -> Unit = {},
-) {
-    var searchText by remember { mutableStateOf("") }
-    var selectedId by remember { mutableIntStateOf(-1) }
-
-    val focusManager = LocalFocusManager.current
-
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(MaterialTheme.colorScheme.onPrimary),
-    ) {
-        AppBarTop(
-            title = stringResource(R.string.select_industry),
-            back = ActionBack(
-                isView = true,
-                onClick = onBack
-            )
-        )
-
-        Surface(
-            color = MaterialTheme.colorScheme.surfaceVariant,
-            shape = RoundedCornerShape(12.dp),
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 16.dp, vertical = 8.dp)
-        ) {
-            OutlinedTextField(
-                value = searchText,
-                onValueChange = { newText ->
-                    searchText = newText
-                    // viewModel.onSearchTextChanged(newText)
-                },
-                placeholder = {
-                    Text(
-                        stringResource(R.string.enter_industry),
-                        color = MaterialTheme.colorScheme.secondary,
-                        style = MaterialTheme.typography.bodyLarge,
-                    )
-                },
-                trailingIcon = {
-                    if (searchText.isEmpty()) {
-                        Icon(
-                            painter = painterResource(R.drawable.ic_search_24),
-                            contentDescription = null,
-                            tint = MaterialTheme.colorScheme.onSurfaceVariant
-                        )
-                    } else {
-                        IconButton(onClick = {
-                            searchText = ""
-                            focusManager.clearFocus()
-                        }) {
-                            Icon(
-                                painter = painterResource(R.drawable.ic_close_24),
-                                contentDescription = "Clear",
-                                tint = MaterialTheme.colorScheme.onSurfaceVariant
-                            )
-                        }
-                    }
-                },
-
-                keyboardOptions = KeyboardOptions(imeAction = ImeAction.Search),
-                keyboardActions = KeyboardActions(onSearch = { focusManager.clearFocus() }),
-                singleLine = true,
-                colors = OutlinedTextFieldDefaults.colors(
-                    focusedBorderColor = MaterialTheme.colorScheme.surfaceVariant,
-                    unfocusedBorderColor = Color.Transparent,
-                    cursorColor = MaterialTheme.colorScheme.primary
-                )
-            )
-        }
-
-        LazyColumn(
-            modifier = Modifier.fillMaxSize()
-        ) {
-            items(industries, key = { it.id }) { industry ->
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .selectableGroup()
-                        .padding(start = 16.dp, bottom = 8.dp, top = 8.dp),
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Text(
-                        text = industry.name,
-                        style = MaterialTheme.typography.bodyLarge,
-                        modifier = Modifier.weight(1f)
-                    )
-                    RadioButton(
-                        selected = selectedId == industry.id,
-                        onClick = {},
-                        colors = RadioButtonDefaults.colors(
-                            selectedColor = MaterialTheme.colorScheme.primary,
-                            unselectedColor = MaterialTheme.colorScheme.primary
-                        )
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun getListIndustry(): List<Industry> {
-    return listOf(
-        Industry(1, "Авиаперевозки"),
-        Industry(2, "Авиационная, вертолетная и аэрокосмическая промышленность"),
-        Industry(4, "Автокомпоненты, запчасти (производство)"),
-        Industry(5, "Автокомпоненты, запчасти, шины (продвижение, оптовая торговля)"),
-        Industry(6, "Автомобильные перевозки"),
-        Industry(7, "Автошкола"),
-        Industry(8, "Агентские услуги в недвижимости"),
-        Industry(9, "Агрохимия (продвижение, оптовая торговля)"),
-        Industry(10, "Агрохимия (производство)"),
-        Industry(11, "Алкогольные напитки (продвижение, оптовая торговля)"),
-        Industry(12, "IT"),
-    )
-}
-
-@Preview(showSystemUi = true, showBackground = true)
-@Composable
-private fun IndustryFilterScreenPreview() {
-    IndustryFilterScreen(
-        industries = getListIndustry(),
-        onBack = {},
-    )
-}
