+++ app/src/main/java/ru/practicum/android/diploma/ui/navigation/ProjectNavHost.kt
@@ -7,10 +7,10 @@ import androidx.navigation.NavType
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.navArgument
+import ru.practicum.android.diploma.filter.industry.ui.IndustryScreen
 import ru.practicum.android.diploma.main.ui.screen.SearchScreen
 import ru.practicum.android.diploma.ui.screens.favorites.FavoritesScreen
 import ru.practicum.android.diploma.ui.screens.filter.FilterScreen
-import ru.practicum.android.diploma.ui.screens.filter.industry.IndustryScreen
 import ru.practicum.android.diploma.ui.screens.filter.workplace.WorkPlaceScreen
 import ru.practicum.android.diploma.ui.screens.team.TeamScreen
 import ru.practicum.android.diploma.vacancy.ui.VacancyDetailScreen
@@ -56,7 +56,9 @@ fun ProjectNavHost(
         }
 
         composable(Routes.INDUSTRY) {
-            IndustryScreen()
+            IndustryScreen(
+                onBack = { navController.popBackStack() },
+            )
         }
 
         composable(Routes.WORKPLACE) {
