+++ app/src/main/res/values/strings.xml
@@ -28,7 +28,6 @@
     <string name="apply">Применить</string>
     <string name="reset">Сбросить</string>
 
-    <string name="select_industries">Выбор отрасли</string>
     <string name="select_work_place">Выбор места работы</string>
     <string name="country">Страна</string>
     <string name="region">Регион</string>
